添加64m和256m存储元件

+++ src/main/java/me/ddggdd135/slimeae/core/slimefun/MECreativeItemStorageCell.java
@@ -12,5 +12,5 @@ public class MECreativeItemStorageCell extends MEItemStorageCell {
     }
 
     @Override
-    public void setSize(int size) {}
+    public void setSize(long size) {}
 }
